Move ObjectFunction to a dedicated file
